✨ :: 학생회 인원 필터 바텀시트 분리
학생회 인원 조회에 쓰던 필터 바텀시트를 member 기능으로 옮기고 전용 선택 모델로 분리했다.
외출 현황 화면이 공용 필터 버튼과 새 멤버 필터 바텀시트를 조합하도록 정리해 책임 경계를 명확하게 맞췄다.

Constraint: 기존 학생회 인원 필터 동작과 선택값 매핑을 유지해야 함
Rejected: 홈 shared 경로에 필터 바텀시트 유지 | member 기능 전용 책임과 재사용 경계가 불명확함
Confidence: high
Scope-risk: narrow
Reversibility: clean
Directive: 학생/학생회 인원 필터 UI는 member_filter_bottom_sheet를 기준으로 확장할 것
Tested: Staged diff review
Not-tested: flutter analyze, screen interaction test
